More optimizations
added several tweaks to hopefully make the script more robust. Some errors might come up due to the way some people name there files. I will be working on a way so some kind of dictionary that users can contribute too so that if the script sees that unknown name it can read the dictionary and get a pointer to what it actually is.